recleaned nLTE files
