+++ radiative_transfer/NGC253HR_nLTE_modelresults.py
@@ -32,23 +32,9 @@ load_observed_results = True
 check_model = False
 spec_index = False
 print_fluxes_latex = False
-
-
-
-
-temp_profiles = False
+models_calculations = True
 fit_logn = False
-plot_models = False
-plot_models_v2 = True # SHC13
-SHC3_plot = False
-mod_difs = False
-
-model_summary = False
-
-models_absorption = False
-comp_LTEvsNLTE = False
-comp_LTEvsNLTE_v15 = False
-
+plot_cont = False
 
 # Cubes
 D_Mpc = 3.5 
@@ -304,389 +290,8 @@ dviolet      = sns.xkcd_palette(['light magenta'])[0]
 dazure      = sns.xkcd_palette(['bright blue'])[0]
 dgreen      = sns.xkcd_palette(['darkish green'])[0]
 
-
-
-        
-if plot_models:
-    Rcrit = 0.85
-    sbagnmodsbestv2= { 'model2': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 1.0, 1.9]],
-                      'model5': ['agn4_LTHC3Nagnsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_a6','dustagnlum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, dviolet, [1, 1, 1.0]],
-                      'model6': ['agn5_LTHC3Nagnsig1.3E+07cd1.0E+25q1.0nsh30rad1.5vt5_a2','dustagnlum1.2E+10cd1.0E+25exp1.0nsh1002rad17',
-                            1.5, dazure, [1, 1, 1.0]],
-                      'model7': ['agn9_LTHC3Nagnsig1.3E+07cd5.6E+24q1.0nsh30rad1.5vt5_x4','dustagnlum1.2E+10cd5.6E+24exp1.0nsh564rad17',
-                            1.5, dgreen, [1, 1, 1.0]],
-               }
     
-    bestmodsv2 = {
-             'model1': ['m13_LTHC3Nsbsig1.3E+07cd1.0E+25q1.0nsh30rad1.5vt5_b3','dustsblum1.2E+10cd1.0E+25exp1.0nsh1002rad17',
-                            1.5, azure, [1, 1, 2.0]],
-             'model2': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 2.3, 1.9]],
-             'model3': ['m23_LTHC3Nsbsig5.5E+07cd1.0E+25q1.0nsh30rad1.5vt5_b7','dustsblum5.0E+10cd1.0E+25exp1.0nsh1002rad17',
-                            1.5, green, [1, 1, 1]],
-             'model4': ['m24_LTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt5_a7','dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, violet, [1, 1, 1]] }
-             
-    modelos = sbagnmodsbestv2
-    # Plot model set
-    
-    #utiles_nLTEmodel.plot_models_and_inp(line_column, modelos, new_hb_df, cont_df, my_model_path, figmod_path, figrt_path, fort_paths)
-    
-    # Final figures
-    #utiles_nLTEmodel.plot_models_and_inp_finalfig(Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths)
-
-    # Final figures present
-    utiles_nLTEmodel.plot_models_and_inp_finalfig_diap(Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths)
-    
-    
-
-    for i,modelo in enumerate(modelos):
-        # modelos[modelo][3] = redpink
-        modsing = {modelo: modelos[modelo]}
-        #utiles_nLTEmodel.line_profiles_chi2(new_hb_df, line_column, modsing, fort_paths, my_model_path, Rcrit)
-        #utiles_nLTEmodel.plot_models_and_inp(Rcrit, line_column, modsing, new_hb_df, cont_df, my_model_path, figmod_path, figrt_path, fort_paths)
-
-
-if plot_models_v2:
-    Rcrit = 0.85
-    sbagnmodsbestv2= { 'model2': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 1.0, 1.9]],
-                      'model5': ['agn4_LTHC3Nagnsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_a6','dustagnlum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, dviolet, [1, 1, 1.0]],
-                      'model6': ['agn5_LTHC3Nagnsig1.3E+07cd1.0E+25q1.0nsh30rad1.5vt5_a2','dustagnlum1.2E+10cd1.0E+25exp1.0nsh1002rad17',
-                            1.5, dazure, [1, 1, 1.0]],
-                      'model7': ['agn9_LTHC3Nagnsig1.3E+07cd5.6E+24q1.0nsh30rad1.5vt5_x4','dustagnlum1.2E+10cd5.6E+24exp1.0nsh564rad17',
-                            1.5, dgreen, [1, 1, 1.0]],
-               }
-    
-    bestmodsv2 = {
-             'model1': ['m13_LTHC3Nsbsig1.3E+07cd1.0E+25q1.0nsh30rad1.5vt5_b3','dustsblum1.2E+10cd1.0E+25exp1.0nsh1002rad17',
-                            1.5, azure, [1, 1, 2.0]],
-             'model2': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 2.3, 1.9]],
-             'model3': ['m23_LTHC3Nsbsig5.5E+07cd1.0E+25q1.0nsh30rad1.5vt5_b7','dustsblum5.0E+10cd1.0E+25exp1.0nsh1002rad17',
-                            1.5, green, [1, 1, 1]],
-             'model4': ['m24_LTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt5_a7','dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, violet, [1, 1, 1]] 
-             }
-             
-    modelos = bestmodsv2
-    # Plot model set
-    
-    #utiles_nLTEmodel.plot_models_and_inp(line_column, modelos, new_hb_df, cont_df, my_model_path, figmod_path, figrt_path, fort_paths)
-    
-    # Final figures
-    #utiles_nLTEmodel.plot_models_and_inp_finalfig(Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths)
-
-    # Final figures present
-    #utiles_nLTEmodel.plot_models_and_inp_finalfig_diap(Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths)
-    
-    # Final paper figures
-    convolve = True # Using or not  convolved fluxes (for adjacent rings)
-    utiles_nLTEmodel.plot_models_and_inp_finalpaperfig(convolve, Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths)
-
-    
-
-    for i,modelo in enumerate(modelos):
-        # modelos[modelo][3] = redpink
-        modsing = {modelo: modelos[modelo]}
-        #utiles_nLTEmodel.line_profiles_chi2(new_hb_df, line_column, modsing, fort_paths, my_model_path, Rcrit)
-        #utiles_nLTEmodel.plot_models_and_inp(Rcrit, line_column, modsing, new_hb_df, cont_df, my_model_path, figmod_path, figrt_path, fort_paths)
-
-if SHC3_plot:
-    finalfigmod_path = my_model_figs+'Models_final_SHC3/'
-    if not os.path.exists(finalfigmod_path):
-        os.makedirs(finalfigmod_path)
-    Rcrit = 0.35
-    rad_pc = 0.50
-    bestmodsv2 = {
-             'model1': ['m13_LTHC3Nsbsig1.3E+07cd1.0E+25q1.0nsh30rad1.5vt5_b3','dustsblum1.2E+10cd1.0E+25exp1.0nsh1002rad17',
-                            1.5, azure, [1, 1, 2.0]],
-             'model2': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 2.3, 1.9]],
-             'model3': ['m23_LTHC3Nsbsig5.5E+07cd1.0E+25q1.0nsh30rad1.5vt5_b7','dustsblum5.0E+10cd1.0E+25exp1.0nsh1002rad17',
-                            1.5, green, [1, 1, 1]],
-             'model4': ['m24_LTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt5_a7','dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, violet, [1, 1, 1]] }
-             
-    modelos = bestmodsv2
-    utiles_nLTEmodel.plot_models_and_inp_finalpaperfig(Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths, source_rad=rad_pc)
-
-
-if models_absorption:
-    Rcrit = 0.85
-    bestmod = {'model2': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 2.3, 1.9]]}
-    modelo = 'm28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9'
-    #utiles_nLTEmodel.read_model_input(bestmod['model2'], my_model_path, Rcrit)
-    #mod_dict = utiles_nLTEmodel.model_summary(modelo, dustmod, my_model_path, distance_pc, Rcrit)
-    
-    
-    
-
-    model2_list = glob(my_model_path+'m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9_*.inp')
-    dustmod = 'dustsblum1.2E+10cd1.0E+25exp1.0nsh1002rad17'
-    rad = 1.5
-    modabs_dict = {}
-    import matplotlib.cm as cm
-    colors = cm.rainbow(np.linspace(0, 1, len(model2_list)))
-    colors = cm.gist_rainbow(np.linspace(0, 1, len(model2_list)))
-    cmap_name = 'my_list'
-    colorlist = [(0.847, 0.057, 0.057),
-                 (0.527, 0.527, 0),
-                 (0, 0.592, 0),
-                 (0, 0.559, 0.559),
-                 (0.316, 0.316, 0.991),
-                 (0.718, 0, 0.718)]
-    from matplotlib.colors import LinearSegmentedColormap
-    cmap = LinearSegmentedColormap.from_list(cmap_name, colorlist, N= len(model2_list))
-    
-    sort_dict = {}
-    for m,mod in enumerate(model2_list):
-        modelo = mod.split('/')[-1].split('.inp')[0]
-        mname = modelo.split('_')[-1]
-        #if float(mname)<=25:
-        #    sort_dict[mname] = modelo
-        sort_dict[mname] = modelo
-    colors = cm.gist_rainbow(np.linspace(0, 1, len(sort_dict)))
-    colors_r = colors#[::-1]
-    for m,mod in enumerate(sort_dict):
-        #modelo = mod.split('/')[-1].split('.inp')[0]
-        #mname = modelo.split('_')[-1]
-        modelo = sort_dict[str(m)]
-        modabs_dict[str(m)] = [modelo, dustmod, rad, colors_r[m], [1, 2.3, 1.9]]
-        
-    modelos = modabs_dict
-    utiles_nLTEmodel.plot_models_and_inp_abscompfig(Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths)
-    
-        
-if comp_LTEvsNLTE:
-    
-    finalfigmod_path = my_model_figs+'Modelscomp_final/'
-    if not os.path.exists(finalfigmod_path):
-        os.makedirs(finalfigmod_path)
-        
-    new_hb_df['v=0_24_23_SM_mJy_kms_beam_345_errcont'] = 1*new_hb_df['v=0_26_25_SM_mJy_kms_beam_345_errcont']
-    new_hb_df['v=0_24_23_SM_mJy_kms_beam_orig_errcont'] = 1*new_hb_df['v=0_26_25_SM_mJy_kms_beam_orig_errcont']
-    #24 26 111 117 336 340
-    line_column = {'plot_conts': [],
-               'v=0_24_23_SM': [15, r'$v=0$  26 - 25', [-10, 150], 1], 
-               'v=0_26_25_SM': [15, r'$v=0$  26 - 25', [-10, 150], 2], 
-               'v7=1_24_1_23_-1_SM': [15, r'$v_{7}=1$  24,1 - 23,-1', [-10, 150], 3],
-               'v7=1_26_1_25_-1_SM': [15, r'$v_{7}=1$  26,1 - 25,-1', [-10, 150], 4],
-               'v6=1_24_-1_23_1_SM': [15, r'$v_{6}=1$  24,-1 - 23,1', [-10, 74], 5],
-               'v6=1_26_-1_25_1_SM': [15, r'$v_{6}=1$  26,-1 - 25,1', [-10, 74], 6],
-               }
-    
-    Rcrit = 0.85
-    compmods1= { 'LTE 3E17': ['m28_cLTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 1.0, 1.9], '-', True],
-                'LTE 3E16': ['m28_cLTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9_low','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, dviolet, [1, 1.0, 1.9], '-', True],
-                'LTE2 3E16': ['mc_cLTHC3Nsbsig1.1E+08cd1.0E+25q1.5nsh30rad1.5vt5_b9_low','dustsblum1.0E+11cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, green, [1, 1.0, 1.9], '-', True],
-                'LTE3 1E16': ['mc_cLTHC3Nsbsig1.1E+08cd1.0E+25q1.5nsh30rad1.5vt5_b9_vlow','dustsblum1.0E+11cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, dgreen, [1, 1.0, 1.9], '-', True],
-                'NLTE 3E16': ['m28_NLTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustagnlum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, violet, [1, 1.0, 1.9], '--', False],
-                'NLTE2 3E16': ['mc_NLTHC3Nsbsig1.1E+08cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.0E+11cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, dazure, [1, 1.0, 1.9], '--', False],
-                'NLTE3 1E16': ['mc_NLTHC3Nsbsig1.1E+08cd1.0E+25q1.5nsh30rad1.5vt5_b9_vlow','dustsblum1.0E+11cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, 'cyan', [1, 1.0, 1.9], '--', False],
-               }
-    
-    
-    compmods= { 'LTE 3E17': ['m28_cLTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 1.0, 1.9], '-', True],
-                'NLTE 3E17': ['m28_NLTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9_up','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, azure, [1, 1.0, 1.9], '--', False],
-                'NLTE 1E16': ['m28_NLTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9_low','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, green, [1, 1.0, 1.9], '--', False],
-               }
-             
-    modelos = compmods
-    # Plot model set
-    
-    #utiles_nLTEmodel.plot_models_and_inp(line_column, modelos, new_hb_df, cont_df, my_model_path, figmod_path, figrt_path, fort_paths)
-    
-    # Comp in figures
-    utiles_nLTEmodel.plot_models_and_inp_comp(Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths)
-    
-    
-    slines = ['ring', 'dist',
-              'v=0_24_23_SM_mJy_kms_beam_orig',
-              'v=0_26_25_SM_mJy_kms_beam_orig',
-              'v7=1_24_1_23_-1_SM_mJy_kms_beam_orig',
-              'v7=1_26_1_25_-1_SM_mJy_kms_beam_orig',
-              'v6=1_24_-1_23_1_SM_mJy_kms_beam_orig',
-              'v6=1_26_-1_25_1_SM_mJy_kms_beam_orig']
-    
-    save_obs_df = new_hb_df[slines]
-    save_obs_df.to_excel(finalfigmod_path+'Observed_fluxes.xlsx')
-    # Final figures present
-    #utiles_nLTEmodel.plot_models_and_inp_finalfig_diap(Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths)
-    
-    modelos = compmods
-    Rcrit = 0.85 # pc Size of the free-free region (where 235/345 starts to rise)
-    redo_lum = False
-    if redo_lum:
-        distance_pc = D_Mpc*(1*u.Mpc).to(u.pc).value
-        modsum_df = pd.DataFrame()
-        for mod in modelos:
-            modelo = modelos[mod][0]
-            dustmod = modelos[mod][1]
-            mod_dict = utiles_nLTEmodel.model_summary(modelo, dustmod, my_model_path, distance_pc, Rcrit)
-            modsum_df = modsum_df.append(mod_dict, ignore_index=True)
-        modsum_df.to_excel(finalfigmod_path+SHC+'_finalmodscomp_modsummary_Rcrit'+str(Rcrit).replace('.','p')+'.xlsx')  
-        
-        
-        
-   
-if comp_LTEvsNLTE_v15:
-    
-    finalfigmod_path = my_model_figs+'Modelscomp_final_v15/'
-    if not os.path.exists(finalfigmod_path):
-        os.makedirs(finalfigmod_path)
-        
-    new_hb_df['v=0_24_23_SM_mJy_kms_beam_345_errcont'] = 1*new_hb_df['v=0_26_25_SM_mJy_kms_beam_345_errcont']
-    new_hb_df['v=0_24_23_SM_mJy_kms_beam_orig_errcont'] = 1*new_hb_df['v=0_26_25_SM_mJy_kms_beam_orig_errcont']
-    #24 26 111 117 336 340
-    line_column = {'plot_conts': [],
-               'v=0_24_23_SM': [15, r'$v=0$  26 - 25', [-10, 150], 1], 
-               'v=0_26_25_SM': [15, r'$v=0$  26 - 25', [-10, 150], 2], 
-               'v7=1_24_1_23_-1_SM': [15, r'$v_{7}=1$  24,1 - 23,-1', [-10, 150], 3],
-               'v7=1_26_1_25_-1_SM': [15, r'$v_{7}=1$  26,1 - 25,-1', [-10, 150], 4],
-               'v6=1_24_-1_23_1_SM': [15, r'$v_{6}=1$  24,-1 - 23,1', [-10, 74], 5],
-               'v6=1_26_-1_25_1_SM': [15, r'$v_{6}=1$  26,-1 - 25,1', [-10, 74], 6],
-               }
-    
-    Rcrit = 0.85
-    compmods1= { 'LTE v5': ['m28_cLTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 1.0, 1.9], '-', True],
-                'LTE v15': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt15_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, azure, [1, 1.0, 1.9], '-', True],
-                'LTE v15c3': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt15_b9_c3','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, azure, [1, 1.0, 1.9], '--', True],
-                'LTE v15c4': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt15_b9_c4','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, 'b', [1, 1.0, 1.9], '--', True],
-                'LTE v15c5': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt15_b9_c5','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink, [1, 1.0, 1.9], '--', True],
-                'LTE v10a3': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt10_b9_a3','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5,  'lime', [1, 1.0, 1.9], '-', True],
-                'LTE v10v20c1': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt10_b9_c1','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5,  green, [1, 1.0, 1.9], '-', True],
-                'LTE v10v20c2': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt10_b9_c2','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5,  green, [1, 1.0, 1.9], '-.', True],
-                'LTE v7a1': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt7_b9_a1','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5,  violet, [1, 1.0, 1.9], '-', True],         
-                'LTE v7a2': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt7_b9_a2','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5,  violet, [1, 1.0, 1.9], '--', True],        
-                'LTE v7c2': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt7_b9_c2','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5,  'k', [1, 1.0, 1.9], '-', True],   
-                'LTE v7c3': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt7_b9_c4','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5,  'k', [1, 1.0, 1.9], '--', True],     
-                                 
-
-               }
-    
-    compmods2= { 'LTE v5': ['m28_cLTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 1.0, 1.9], '-', True],
-                'LTE HL v15': ['m24_LTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt15_a7','dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, azure, [1, 1.0, 1.9], '-', True],
-                               
-                'NLTE HL v15': ['m24_NLTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt15_a7', 'dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, green, [1, 1.0, 1.9], '--', False],
-                'NLTE HL v10v15': ['m24_NLTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt15_a7_c3', 'dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, dgreen, [1, 1.0, 1.9], '--', False],
-                'NLTE HL v10v15v20': ['m24_NLTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt15_a7_c5', 'dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, violet, [1, 1.0, 1.9], '--', False],
-                'NLTE HL v10v15v20c': ['m24_NLTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt15_a7_c6', 'dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, dviolet, [1, 1.0, 1.9], '--', False],
-                'NLTE HL v10v15v20d': ['m24_NLTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt15_a7_c7', 'dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, 'k', [1, 1.0, 1.9], '--', False],
-             }
-                
-    
-    #'model4': ['m24_LTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt5_a7','dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-    #                        1.5, violet, [1, 1, 1]] }
-
-    modelos = compmods1
-    # Plot model set
-    
-    #utiles_nLTEmodel.plot_models_and_inp(line_column, modelos, new_hb_df, cont_df, my_model_path, figmod_path, figrt_path, fort_paths)
-    
-    # Comp in figures
-    utiles_nLTEmodel.plot_models_and_inp_comp(Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths)
-    
-    
-    slines = ['ring', 'dist',
-              'v=0_24_23_SM_mJy_kms_beam_orig',
-              'v=0_26_25_SM_mJy_kms_beam_orig',
-              'v7=1_24_1_23_-1_SM_mJy_kms_beam_orig',
-              'v7=1_26_1_25_-1_SM_mJy_kms_beam_orig',
-              'v6=1_24_-1_23_1_SM_mJy_kms_beam_orig',
-              'v6=1_26_-1_25_1_SM_mJy_kms_beam_orig']
-    
-    save_obs_df = new_hb_df[slines]
-    save_obs_df.to_excel(finalfigmod_path+'Observed_fluxes.xlsx')
-    # Final figures present
-    #utiles_nLTEmodel.plot_models_and_inp_finalfig_diap(Rcrit, line_column, modelos, new_hb_df, cont_df, my_model_path, finalfigmod_path, finalfigmod_path, fort_paths)
-    
-    modelos = compmods
-    Rcrit = 0.85 # pc Size of the free-free region (where 235/345 starts to rise)
-    redo_lum = False
-    if redo_lum:
-        distance_pc = D_Mpc*(1*u.Mpc).to(u.pc).value
-        modsum_df = pd.DataFrame()
-        for mod in modelos:
-            modelo = modelos[mod][0]
-            dustmod = modelos[mod][1]
-            mod_dict = utiles_nLTEmodel.model_summary(modelo, dustmod, my_model_path, distance_pc, Rcrit)
-            modsum_df = modsum_df.append(mod_dict, ignore_index=True)
-        modsum_df.to_excel(finalfigmod_path+SHC+'_finalmodscomp_modsummary_Rcrit'+str(Rcrit).replace('.','p')+'.xlsx')  
-        
-    
-if model_summary: 
-    
-    allbestmodsv2 = {
-             'model1': ['m13_LTHC3Nsbsig1.3E+07cd1.0E+25q1.0nsh30rad1.5vt5_b3','dustsblum1.2E+10cd1.0E+25exp1.0nsh1002rad17',
-                            1.5, azure, [1, 1, 2.0]],
-             'model2': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b9','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                            1.5, redpink , [1, 2.3, 1.9]],
-             'model3': ['m23_LTHC3Nsbsig5.5E+07cd1.0E+25q1.0nsh30rad1.5vt5_b7','dustsblum5.0E+10cd1.0E+25exp1.0nsh1002rad17',
-                            1.5, green, [1, 1, 1]],
-             'model4': ['m24_LTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt5_a7','dustsblum5.0E+10cd1.0E+25exp1.5nsh1003rad17',
-                    1.5, violet, [1, 1, 1]],
-             'model5': ['agn4_LTHC3Nagnsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_a6','dustagnlum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                    1.5, dviolet, [1, 1, 1.0]],
-             'model6': ['agn5_LTHC3Nagnsig1.3E+07cd1.0E+25q1.0nsh30rad1.5vt5_a2','dustagnlum1.2E+10cd1.0E+25exp1.0nsh1002rad17',
-                    1.5, dazure, [1, 1, 1.0]],
-             'model7': ['agn9_LTHC3Nagnsig1.3E+07cd5.6E+24q1.0nsh30rad1.5vt5_x4','dustagnlum1.2E+10cd5.6E+24exp1.0nsh564rad17',
-                    1.5, dgreen, [1, 1, 1.0]]
-                      }
-
-    modelos = allbestmodsv2
-    Rcrit = 0.85 # pc Size of the free-free region (where 235/345 starts to rise)
-    redo_lum = False
-    if redo_lum:
-        distance_pc = D_Mpc*(1*u.Mpc).to(u.pc).value
-        modsum_df = pd.DataFrame()
-        for mod in modelos:
-            modelo = modelos[mod][0]
-            dustmod = modelos[mod][1]
-            mod_dict = utiles_nLTEmodel.model_summary(modelo, dustmod, my_model_path, distance_pc, Rcrit)
-            modsum_df = modsum_df.append(mod_dict, ignore_index=True)
-        modsum_df.to_excel(final_results_path+SHC+'_finalmods_modsummary_Rcrit'+str(Rcrit).replace('.','p')+'.xlsx')  
-        
-        
-        
-        
-    modsum_df_rcrit0p5 = pd.read_excel(final_results_path+SHC+'_finalmods_modsummary_Rcrit0p5.xlsx', header=0)
-    
-    
-        
+if models_calculations:  
     distance_pc = 3.5e6
     # Reading saved model summary
     TcenAGN = 1300 # Temperatural central del AGN
@@ -1213,12 +818,6 @@ if model_summary:
         fig.savefig(final_results_path+'LIRMgas_Lir.pdf', bbox_inches='tight', transparent=True, dpi=400)
         plt.close()
         
-        
-        
-        
-        
-    
-        
 if fit_logn:
     mname = 'm24_LTHC3Nsbsig5.5E+07cd1.0E+25q1.5nsh30rad1.5vt5_a4'
     fit_XHC3N, x_profile, shell_dists, model_dists = utiles_nLTEmodel.plot_model_input([mname], my_model_path, figinp_path)
@@ -1232,111 +831,6 @@ if fit_logn:
     for xhc3n in x_profile:
         print(f'{xhc3n:1.1e}')
 
-
-
-if mod_difs:
-     
-    for m, mod in enumerate(bestmod):
-        
-        SM_chanwidth_kms_230GHz = 2.49
-        SM_chanwidth_kms_218GHz = 2.67
-        SM_chanwidth_kms_220GHz = 2.65 
-        # Errors with cont
-        for l,line in enumerate(line_column):
-            if line not in ['plot_conts', 'plot_T', 'plot_col', 'plot_x']:
-                hb_df[line+'_mJy_kms_beam_orig_errcont'] = np.sqrt(hb_df[line+'_mJy_kms_beam_orig_err']**2 + hb_df['cont_'+line+'_beam_orig_err']**2)
-                hb_df[line+'_mJy_kms_beam_345_errcont'] = np.sqrt(hb_df[line+'_mJy_kms_beam_345_err']**2 + hb_df['cont_'+line+'_beam_345_err']**2)
-                hb_df[line+'_width_kms'] = np.abs(hb_df[line+'_vmin']-hb_df[line+'_vmax'])
-                hb_df[line+'_width_channs'] = hb_df[line+'_width_kms']/SM_chanwidth_kms_230GHz
-        
-        
-        bestmod = {'mod28b1': ['m28_LTHC3Nsbsig1.3E+07cd1.0E+25q1.5nsh30rad1.5vt5_b1','dustsblum1.2E+10cd1.0E+25exp1.5nsh1003rad17',
-                                    1.5, redpink , [1, 1, 1]]}
-        
-        modelo = bestmod[mod]
-        mdust, m_molec, m_molec345 = utiles_nLTEmodel.model_reader(modelo, fort_paths, 1, 1, 1, line_column, my_model_path, LTE=True, read_only=True)
-        lineas = {
-                   'v7=1_24_1_23_-1_SM': [15, r'$v_{7}=1$  24,1 - 23,-1', [-10, 130], 6],
-                   'v7=1_26_1_25_-1_SM': [15, r'$v_{7}=1$  26,1 - 25,-1', [-10, 130], 2],
-                   'v6=1_24_-1_23_1_SM': [15, r'$v_{6}=1$  24,-1 - 23,1', [-10, 95], 10],
-                   'v6=1_26_-1_25_1_SM': [15, r'$v_{6}=1$  26,-1 - 25,1', [-10, 95], 3],
-                   'v7=2_24_0_23_0_SM':  [15, r'$v_{7}=2$  24,0 - 23,0', [-5, 70], 11],
-                   'v7=2_26_0_25_0_SM':  [15, r'$v_{7}=2$  26,0 - 25,0', [-5, 70], 4],
-                   'v5=1_v7=3_26_1_0_25_-1_0_SM':  [10, r'$v_{5}=1/v_7=3$  26,1,0 - 25,-1,0', [-3, 40], 7],
-                   'v6=v7=1_26_2_25_-2_SM': [10, r'$v_{6}=v_{7}=1$  26,2,2 - 25,-2,2', [-3, 30],  8],
-                   'v=0_26_25_SM': [15, r'$v=0$  26 - 25', [-10, 150], 1],
-                   'v4=1_26_25_SM': [10, r'$v_{4}=1$  26 - 25', [-3, 30], 5],
-                   'v6=2_24_0_23_0_SM':  [10, r'$v_{6}=2$  24,0 - 23,0', [-3, 30], 9],
-                   }
-        
-        maxis = 4
-        naxis = 3
-        figsize = 20
-        fontsize = 12
-        fig = plt.figure(figsize=(figsize*naxis/maxis*1.15, figsize*0.85))
-        gs1 = gridspec.GridSpec(maxis, naxis)#, width_ratios=[1,1,1,0.1], height_ratios=[1])    
-        gs1.update(wspace = 0.1, hspace=0.0, top=0.95, bottom = 0.05, left=0.05, right=0.80)
-        axes=[]
-        
-        
-        for l,line in enumerate(lineas):
-            poly = np.polyfit(hb_df['dist'], hb_df[line+'_mJy_kms_beam_orig'], deg=10)
-            fit_profile= np.polyval(poly, m_molec[0])
-            axes.append(fig.add_subplot(gs1[l]))
-            
-            axes[l].plot(new_hb_df['dist'], new_hb_df[line+'_mJy_kms_beam_orig'], marker='s', linestyle='', color='b',zorder=2)
-            axes[l].plot(new_hb_df['dist'], new_hb_df[line+'_mJy_kms_beam_345'], marker='s', linestyle='', color='g',zorder=2)
-            
-            axes[l].plot(hb_df['dist'], hb_df[line+'_mJy_kms_beam_orig'], marker='o', linestyle='', color='k',zorder=2)
-            axes[l].plot(hb_df['dist'], hb_df[line+'_mJy_kms_beam_345'], marker='o', linestyle='', color='r',zorder=2)
-            axes[l].plot(m_molec[0], m_molec[line+'_beam_orig'], linestyle='-', markersize=5, color='b')
-            if mod_difs:
-                diff = []
-                axes[l].plot(m_molec[0], fit_profile, linestyle='-', markersize=5, color='r')
-                for r, ring in hb_df.iterrows():
-                    
-                    for d, dist in enumerate(m_molec[0]):
-                        if dist == ring['dist']:
-                            difer = np.abs(m_molec[line+'_beam_orig'][d]-ring[line+'_mJy_kms_beam_orig'])
-                            diff.append(difer)
-                            hb_df.loc[r, line+'_mJy_kms_beam_orig_difmod'] = difer
-                            print(f' Found dist {dist}')
-                hb_df[line+'_intcont'] = hb_df['cont_'+line+'_beam_orig']*hb_df[line+'_width_kms']
-                hb_df[line+'_contdifmean'] = hb_df[line+'_mJy_kms_beam_orig_difmod']/hb_df[line+'_width_kms']
-            
-            #minor_locator = AutoMinorLocator(2)
-            axes[l].set_xlim([0.0, 1.5])
-            axes[l].set_ylim(lineas[line][2])
-            axes[l].tick_params(direction='in')
-            axes[l].tick_params(axis="both", which='major', length=8)
-            axes[l].tick_params(axis="both", which='minor', length=4)
-            axes[l].xaxis.set_tick_params(which='both', top ='on')
-            axes[l].yaxis.set_tick_params(which='both', right='on', labelright='off')
-            axes[l].tick_params(axis='both', which='major')
-            
-            axes[l].tick_params(labelleft=True,
-                           labelright=False)
-            axes[l].text(0.9, 0.95, lineas[line][1],
-                                horizontalalignment='right',
-                                verticalalignment='top',
-                                fontsize=fontsize,
-                                transform=axes[l].transAxes)
-            
-            if l <= 7:
-                axes[l].tick_params(
-                           labelbottom=False)
-            else:
-                axes[l].set_xlabel(r'r (pc)')
-        fig_spec = NGC253_path+'SHC/'+SHC+'/EdFlux/Figures_python/'
-        fig.savefig(fig_spec+SHC+'_fluxes_comp.pdf', bbox_inches='tight', transparent=True, dpi=400)
-        plt.close()
-
-
-
-
-
-
-plot_cont = False
 if plot_cont:
     NGC253_path = '/Users/frico/Documents/data/NGC253_HR/'
     cont_path = 'Continuums/'
